Update README.md and introducing new features for the next update
New update for the README.md file. Nothing new for the site but we're working hard for a new update that will come in this weekend or next week. This is a really big update, with a brand new navigation-bar and new pages for games. We will also introduce our brand new AI chat-bot: Ary. With this bot it's possible to ask simple questions and we'll introduce more features later!

Stay in touch trough the news-page on our website!
Ryan and Armaan