June Update: AryaLM2.1, LeerToGo and more!
Hello there!

We're excited to tell about our big June Update! We're launching AryaLM2.1, our best and smartest model yet. It can better understand your input and can give better answers. We also bring out LeerToGo, our new tool to create lists and play them in a quiz. We've also solved some bugs in the Register-form and redesigned the Arya-page.

Have fun with this new update and we'll be back in the Summer-holiday with a brand new update!
Ryan and Armaan