AryaLM3.1 Is Here!
[VERSION 3.1]
Welcome everyone!

In this update, we officially launch AryaLM3.1, our newest and best LM ever! This model is exclusively for registered users only, so make sure you have an account to directly use this new model.

And.... do you ever want to train Arya on your own data? That's now possible with the Arya Trainer, our tool to type the input and output (I/O) and instantly get the code for Arya. You can sent this to us through the Contact-page and we'll add this code in the next Language Model of Arya!

For the last part of the update we've solved some bugs and added a more modern design to some pages.

Check out the update on: https://ryanvanderhaak.github.io/Online-Leerhulp/Home.html and we'll see you in the next update!