[3.1.1] Security-update and other improvements
Hello there!

We're releasing a new security-update which improve the security of Arya and the pages for your account. Arya is now only available with a account, so make sure you make one! There is also a limit on the messages; 10 in a minute. This update also brings a feature to reset your password in case you forgot it.

Enjoy this update!