LeerToGo-update, more security and other improvements!
Hello!

We've added new possibilities for LeerToGo to check your answers with your Enter-key. We've also added extra security-protocols to our site. To guarantee the security, we won't say what kind of things we've added. You will experience those things when using the site.

Good luck in this new update!
The Online Leerhulp-team